# docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Онлайн-кинотеатр API"
  description: |
    Полноценный API для стримингового сервиса с поддержкой фильмов, сериалов, рейтингов и пользовательских закладок.
    
    ## Основные возможности:
    - 📺 Просмотр фильмов и сериалов
    - ⭐ Система рейтингов и отзывов
    - 📋 Личные закладки пользователей
    - 👥 Управление пользователями и подписками
    - 📊 Аналитика и статистика
    - 🔧 Управление справочниками (админ)
    
    ## Аутентификация:
    Все защищенные эндпоинты требуют JWT токен, полученный через `/auth/login`.
    
    ## Технологии:
    - FastAPI 0.104+
    - SQLModel (SQLAlchemy + Pydantic)
    - JWT аутентификация
    - PostgreSQL / SQLite

  version: "2.0.0"
  contact:
    name: "API Support"
    email: "support@cinema.example.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://localhost:8000"
    description: "Локальный сервер разработки"
  - url: "https://api.cinema.example.com"
    description: "Продакшн сервер"
...
paths:
...
  /ratings/{content_id}:
    post:
      summary: "Добавить или обновить рейтинг"
      description: "Добавление или обновление оценки контента"
      tags: ["Рейтинги"]
      parameters:
        - name: content_id
          in: path
          required: true
          description: "ID контента"
          schema:
            type: integer
            minimum: 1
        - name: rating_value
          in: query
          required: true
          description: "Оценка от 0.5 до 5"
          schema:
            type: number
            minimum: 0.5
            maximum: 5.0
            multipleOf: 0.5
        - name: review
          in: query
          required: false
          description: "Текстовый отзыв"
          schema:
            type: string
            maxLength: 1000
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Рейтинг успешно добавлен/обновлен"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingOperationResponse"
        "400":
          description: "Неверная оценка"
        "401":
          description: "Неавторизованный доступ"
        "404":
          description: "Контент не найден"

    get:
      summary: "Получить рейтинги контента"
      description: "Получение всех оценок и отзывов для контента"
      tags: ["Рейтинги"]
      parameters:
        - name: content_id
          in: path
          required: true
          description: "ID контента"
          schema:
            type: integer
            minimum: 1
        - name: include_reviews
          in: query
          required: false
          description: "Включать отзывы в ответ"
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          required: false
          description: "Количество элементов на странице"
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: "Смещение для пагинации"
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: "Рейтинги контента"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentRatingsResponse"
        "404":
          description: "Контент не найден"
  /ratings/user/{user_id}:
    get:
      summary: "Получить рейтинги пользователя"
      description: "Получение всех оценок пользователя (только для владельца)"
      tags: ["Рейтинги"]
      parameters:
        - name: user_id
          in: path
          required: true
          description: "ID пользователя"
          schema:
            type: integer
            minimum: 1
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Рейтинги пользователя"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRatingsResponse"
        "401":
          description: "Неавторизованный доступ"
        "403":
          description: "Недостаточно прав"
        "404":
          description: "Пользователь не найден"

  # ============ Закладки ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT токен аутентификации"
  schemas:
...
    RatingOperationResponse:
      type: object
      properties:
        rating:
          $ref: "#/components/schemas/RatingResponse"
        average_rating:
          type: number
          example: 4.5
        total_ratings:
          type: integer
          example: 125
    RatingResponse:
      type: object
      required: [id, user_id, content_id, rating_value, created_at]
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        content_id:
          type: integer
          example: 1
        rating_value:
          type: number
          minimum: 0.5
          maximum: 5.0
          multipleOf: 0.5
          example: 4.5
        review:
          type: string
          nullable: true
          example: "Отличный фильм, рекомендую!"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    RatingsSummary:
      type: object
      properties:
        total_ratings:
          type: integer
          example: 125
        average_rating:
          type: number
          example: 4.5
        min_rating:
          type: number
          example: 1.0
        max_rating:
          type: number
          example: 5.0
    ContentRatingsResponse:
      type: object
      properties:
        content_id:
          type: integer
          example: 1
        content_title:
          type: string
          example: "Интерстеллар"
        statistics:
          $ref: "#/components/schemas/RatingsSummary"
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/RatingResponse"
    UserRatingsResponse:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        total_ratings:
          type: integer
          example: 15
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/RatingResponse"

    # ============ Закладки ============
...